Fixed bug in epoll when joining normal and tssx events

+++ include/tssx/epoll-overrides.h
@@ -51,7 +51,7 @@ typedef struct EpollTask {
 	size_t number_of_events;
 	int timeout;
 	event_count_t *shared_event_count;
-	size_t event_count;
+	int event_count;
 } EpollTask;
 
 /******************** REAL FUNCTIONS ********************/
